added tests for kill node and test_bs refactor
