+++ tests/late_nodes/test_delayed.py
@@ -1,15 +1,18 @@
 import time
+from datetime import datetime, timedelta
 
+import pytz
 from pytest_testconfig import config as test_config
+
 
 from tests import queries
 from tests.conftest import DeploymentInfo
 from tests.deployment import create_deployment
 from tests.hare.assert_hare import expect_hare
-from tests.misc import CoreV1ApiClient
+from tests.misc import CoreV1ApiClient, get_conf
 from tests.queries import query_atx_published
 from tests.test_bs import save_log_on_exit, setup_bootstrap, add_curl, start_poet
-from tests.test_bs import current_index, CLIENT_DEPLOYMENT_FILE, get_conf
+from tests.test_bs import current_index, CLIENT_DEPLOYMENT_FILE
 
 
 def new_client_in_namespace(name_space, setup_bootstrap, cspec, num):
@@ -59,6 +62,7 @@ def sleep_and_print(total_seconds):
 #    TESTS
 # ==============================================================================
 
+GENESIS_TIME = pytz.utc.localize(datetime.utcnow() + timedelta(seconds=test_config['genesis_delta']))
 
 def test_add_delayed_nodes(init_session, add_curl, setup_bootstrap, start_poet, save_log_on_exit):
     bs_info = setup_bootstrap.pods[0]
