+++ tests/p2p/test_p2p.py
@@ -9,9 +9,10 @@ import time
 from pytest_testconfig import config as testconfig
 # noinspection PyUnresolvedReferences
 from tests.context import ES
+from tests.misc import add_multi_clients, get_conf
 
 from tests.queries import query_message, poll_query_message
-from tests.test_bs import api_call, add_multi_clients, get_conf
+from tests.test_bs import api_call
 # noinspection PyUnresolvedReferences
 from tests.test_bs import setup_bootstrap, setup_clients, save_log_on_exit, add_curl
 
@@ -47,7 +48,7 @@ def add_client(request, setup_bootstrap, setup_clients):
 
         bs_info = setup_bootstrap.pods[0]
         cspec = get_conf(bs_info, testconfig['client'])
-        client_name = add_multi_clients(setup_bootstrap.deployment_id, cspec, 1)[0]
+        client_name = add_multi_clients(testconfig, setup_bootstrap.deployment_id, cspec, 1)[0]
         return client_name
 
     return _add_single_client()
@@ -88,7 +89,7 @@ def add_clients(setup_bootstrap, setup_clients):
             client_key += f'{version_separator}{version}'
 
         cspec = get_conf(bs_info, testconfig[client_key])
-        pods_names = add_multi_clients(setup_bootstrap.deployment_id, cspec, size=num_of_clients)
+        pods_names = add_multi_clients(testconfig, setup_bootstrap.deployment_id, cspec, size=num_of_clients)
         return pods_names
 
     return _add_clients
@@ -140,7 +141,7 @@ def test_add_many_clients(init_session, setup_bootstrap, setup_clients):
     bs_info = setup_bootstrap.pods[0]
     cspec = get_conf(bs_info, testconfig['client'])
 
-    pods = add_multi_clients(setup_bootstrap.deployment_id, cspec, size=4)
+    pods = add_multi_clients(testconfig, setup_bootstrap.deployment_id, cspec, size=4)
     time.sleep(40 * timeout_factor)  # wait for the new clients to finish bootstrap and for logs to get to elasticsearch
     fields = {'M': 'discovery_bootstrap'}
     for p in pods:
@@ -343,7 +344,7 @@ def test_late_bootstraps(init_session, setup_bootstrap, setup_clients):
     testnames = []
 
     for i in range(TEST_NUM):
-        client = add_multi_clients(setup_bootstrap.deployment_id,
+        client = add_multi_clients(testconfig, setup_bootstrap.deployment_id,
                                    get_conf(setup_bootstrap.pods[0], testconfig['client']),
                                    1)
         testnames.append((client[0], datetime.now()))
